MT#59962 XMLRPC2DI: `time_t` use long unsigned for printing
Fixies:

*** CID 544159:  API usage errors  (PW.PRINTF_ARG_MISMATCH)
/apps/xmlrpc2di/XMLRPC2DI.cpp: 94 in ()
88
89       bool multithreaded = (cfg.getParameter("multithreaded") == "yes");
90       DebugServerResult = (cfg.getParameter("debug_server_result", "no") == "yes");
91       DebugServerParams = (cfg.getParameter("debug_server_params", "no") == "yes");
92       ServerRetryAfter = cfg.getParameterInt("server_retry_after", 10);
93
>>>     CID 544159:  API usage errors  (PW.PRINTF_ARG_MISMATCH)
>>>     argument is incompatible with corresponding format string conversion (expected type "unsigned int" but argument has type "long")
94       DBG("Parameter server_retry_after is set to '%u' seconds\n", ServerRetryAfter);
95
96       /* threads */
97       if (multithreaded) {
98         threads = cfg.getParameterInt("threads", 5);
99       }

+++ apps/xmlrpc2di/XMLRPC2DI.cpp
@@ -91,7 +91,7 @@ int XMLRPC2DI::load() {
   DebugServerParams = (cfg.getParameter("debug_server_params", "no") == "yes");
   ServerRetryAfter = cfg.getParameterInt("server_retry_after", 10);
 
-  DBG("Parameter server_retry_after is set to '%u' seconds\n", ServerRetryAfter);
+  DBG("Parameter server_retry_after is set to '%lu' seconds\n", static_cast<unsigned long>(ServerRetryAfter));
 
   /* threads */
   if (multithreaded) {
